fix: extract the ChatContext
